# api/openapi.yaml
openapi: 3.1.0
info:
  title: video-record API
  version: 1.0.0
  description: Private movie and TV viewing diary API.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/collections:
    get:
      operationId: listCollections
      summary: List private collections
      responses:
        '200':
          description: Private collections.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CollectionList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createCollection
      summary: Create a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCollectionRequest' }
      responses:
        '201':
          description: Collection created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Collection' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/collections/{collectionID}/items:
    parameters:
      - name: collectionID
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: addCollectionItem
      summary: Add media to a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddCollectionItemRequest' }
      responses:
        '204': { description: Collection item added. }
        default: { $ref: '#/components/responses/Problem' }
    put:
      operationId: replaceCollectionItems
      summary: Replace and order items in a private collection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceCollectionItemsRequest' }
      responses:
        '204': { description: Collection items replaced in the supplied order. }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/records/{mediaID}:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getRecord
      summary: Read the current user's media record
      responses:
        '200':
          description: Current media record.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordState' }
        default: { $ref: '#/components/responses/Problem' }
    put:
      operationId: updateRecord
      summary: Update a versioned current-user media record
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRecordRequest' }
      responses:
        '200':
          description: Updated media record.
          headers:
            ETag: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecordState' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/records/{mediaID}/events:
    parameters:
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: listWatchEvents
      summary: List immutable watch events for a media record
      responses:
        '200':
          description: Watch events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchEventList' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createWatchEvent
      summary: Create an immutable rewatch event
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWatchEventRequest' }
      responses:
        '201':
          description: Watch event created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchEvent' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/records/{mediaID}/events/{eventID}:
    parameters:
      - $ref: '#/components/parameters/MediaID'
      - name: eventID
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteWatchEvent
      summary: Delete an owned watch event
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204': { description: Watch event deleted. }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/household/records/{ownerID}/{mediaID}:
    parameters:
      - name: ownerID
        in: path
        required: true
        schema: { type: string, format: uuid }
      - $ref: '#/components/parameters/MediaID'
    get:
      operationId: getVisibleHouseholdRecord
      summary: Read explicitly shared household record fields
      responses:
        '200':
          description: Explicitly shared record fields.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VisibleRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: video_record_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^"\d+"$' }
    MediaID:
      name: mediaID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    RecordStatus:
      type: string
      enum: [none, wishlist, watching, completed, dropped]
    RecordSource:
      type: string
      enum: [manual, confirmed_import, confirmed_sync, external_default]
...
    Collection:
      type: object
      additionalProperties: false
      required: [id, name, items]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        items:
          type: array
          items: { type: string, format: uuid }
    CollectionList:
      type: array
      items: { $ref: '#/components/schemas/Collection' }
    CreateCollectionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    AddCollectionItemRequest:
      type: object
      additionalProperties: false
      required: [mediaId]
      properties:
        mediaId: { type: string, format: uuid }
    ReplaceCollectionItemsRequest:
      type: object
      additionalProperties: false
      required: [mediaIds]
      properties:
        mediaIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
...
    RecordState:
      type: object
      additionalProperties: false
      required: [mediaId, status, rating, note, watchedAt, viewingMethod, version]
      properties:
        mediaId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RecordStatus' }
        rating: { type: [number, 'null'], minimum: 0, maximum: 10 }
        note: { type: [string, 'null'] }
        watchedAt: { type: [string, 'null'], format: date-time }
        viewingMethod: { type: [string, 'null'] }
        version: { type: integer, minimum: 0 }
    UpdateRecordRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/RecordStatus' }
        rating: { type: [number, 'null'], minimum: 0, maximum: 10 }
        note: { type: [string, 'null'] }
        watchedAt: { type: string, format: date-time }
        viewingMethod: { type: [string, 'null'] }
        participantIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
...
    WatchEvent:
      type: object
      additionalProperties: false
      required: [id, mediaId, watchedAt, source, completion]
      properties:
        id: { type: string, format: uuid }
        mediaId: { type: string, format: uuid }
        episodeId: { type: string, format: uuid }
        watchedAt: { type: string, format: date-time }
        viewingMethod: { type: string }
        source: { $ref: '#/components/schemas/RecordSource' }
        externalEventId: { type: string }
        completion: { type: integer, minimum: 0, maximum: 100 }
        note: { type: string }
    WatchEventList:
      type: array
      items: { $ref: '#/components/schemas/WatchEvent' }
    CreateWatchEventRequest:
      type: object
      additionalProperties: false
      required: [watchedAt]
      properties:
        watchedAt: { type: string, format: date-time }
        viewingMethod: { type: string }
        note: { type: string }
...
    VisibleRecord:
      type: object
      additionalProperties: false
      required: [ownerId, mediaId, rating, privateNote, sharedReview]
      properties:
        ownerId: { type: string, format: uuid }
        mediaId: { type: string, format: uuid }
        rating: { type: [number, 'null'], minimum: 0, maximum: 10 }
        privateNote: { type: [string, 'null'] }
        sharedReview: { type: [string, 'null'] }
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        requestId: { type: string }